Update deDuplicateDBs.py
Add context-dependent messages about how many duplicate entries there are to remove!

+++ programFiles/extra/deDuplicateDBs.py
@@ -20,7 +20,7 @@ def getAllEntries(**args):
     return entries
 
 if Path.cwd().joinpath('places.sqlite').is_file() == False:
-    print('There is no places.sqlite DB in this folder. Please copy this file to here and rerun the program.')
+    print('There is no places.sqlite DB in this folder. Please copy the file here and rerun this program.')
 
 elif Path.cwd().joinpath('places.sqlite').is_file() == True:
     dbMain = sqlite3.connect('places.sqlite')
@@ -31,14 +31,27 @@ elif Path.cwd().joinpath('places.sqlite').is_file() == True:
     dupSQL = 'SELECT from_visit, place_id, visit_date, id, COUNT(*) Num FROM moz_historyvisits group by from_visit, place_id, visit_date having Num > 1'
     duplicateVisits = getAllEntries(cur = curMain, SQL = dupSQL, dictSchema ='(entry[0], entry[1], entry[2]): [entry[3], entry[4]]')
 
+    numDuplicates = 0
+    for dup in duplicateVisits.values():
+        numDuplicates += dup[1]
 
-    done = []
-    for visit in allVisits.values():
-        if (visit[1], visit[2], visit[3]) in duplicateVisits.keys() and (visit[1], visit[2], visit[3]) not in done:
-            lowestID, numDuplicates = duplicateVisits[(visit[1], visit[2], visit[3])][:]
-            curMain.execute('DELETE from main.moz_historyvisits where from_visit = ? and place_id = ? and visit_date = ? and id > ?', (visit[1], visit[2], visit[3], lowestID))
-            done.append((visit[1], visit[2], visit[3]))
+    numDuplicates -= len(duplicateVisits)
 
+    if numDuplicates > 0:
+        print(f'Removing {numDuplicates} duplicate entries...')
+
+        done = []
+        for visit in allVisits.values():
+            if (visit[1], visit[2], visit[3]) in duplicateVisits.keys() and (visit[1], visit[2], visit[3]) not in done:
+                lowestID, numDuplicates = duplicateVisits[(visit[1], visit[2], visit[3])][:]
+                curMain.execute('DELETE from main.moz_historyvisits where from_visit = ? and place_id = ? and visit_date = ? and id > ?', (visit[1], visit[2], visit[3], lowestID))
+                done.append((visit[1], visit[2], visit[3]))
+
+
+        input('\nDone.')
+
+    elif numDuplicates == 0:
+        print('No duplicate entries found.')
 
     curMain.connection.commit()
     curMain.close()
